+++ pueue/src/daemon/network/certificate.rs
@@ -1,4 +1,4 @@
-use std::{fs::File, io::Write, path::Path};
+use std::{fs::OpenOptions, io::Write, path::Path};
 
 use rcgen::{CertifiedKey, generate_simple_self_signed};
 
@@ -30,35 +30,37 @@ pub fn create_certificates(shared_settings: &Shared) -> Result<(), Error> {
         .map_err(|_| {
             Error::CertificateFailure("Failed to generate self-signed daemon certificate.".into())
         })?;
-    // The certificate is now valid for localhost and the domain "hello.world.example"
+    // The certificate is now valid for localhost
     let ca_cert = cert.pem();
-    write_file(ca_cert, "daemon cert", &daemon_cert_path)?;
+    write_file(ca_cert, "daemon cert", &daemon_cert_path, 0o640)?;
 
+    // The key is only ever read by the daemon itself, so it doesn't need to be group readable.
     let ca_key = signing_key.serialize_pem();
-    write_file(ca_key, "daemon key", &daemon_key_path)?;
+    write_file(ca_key, "daemon key", &daemon_key_path, 0o600)?;
 
     Ok(())
 }
 
-fn write_file(blob: String, name: &str, path: &Path) -> Result<(), Error> {
+/// Write a certificate or key with a specific mode.
+#[cfg_attr(target_os = "windows", allow(unused_variables))]
+fn write_file(blob: String, name: &str, path: &Path, mode: u32) -> Result<(), Error> {
     info!("Generate {name}.");
-    let mut file = File::create(path)
+
+    let mut options = OpenOptions::new();
+    options.write(true).create_new(true);
+
+    #[cfg(not(target_os = "windows"))]
+    {
+        use std::os::unix::fs::OpenOptionsExt;
+        options.mode(mode);
+    }
+
+    let mut file = options
+        .open(path)
         .map_err(|err| Error::IoPathError(path.to_path_buf(), "creating certificate", err))?;
 
     file.write_all(&blob.into_bytes())
         .map_err(|err| Error::IoPathError(path.to_path_buf(), "writing certificate", err))?;
-
-    #[cfg(not(target_os = "windows"))]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        let mut permissions = file
-            .metadata()
-            .map_err(|_| Error::CertificateFailure("Failed to certificate permission.".into()))?
-            .permissions();
-        permissions.set_mode(0o640);
-        std::fs::set_permissions(path, permissions)
-            .map_err(|_| Error::CertificateFailure("Failed to certificate permission.".into()))?;
-    }
 
     Ok(())
 }